ab_client_consensus_common

Module consensus_parameters

Structs

DeriveConsensusParametersConsensusInfo

A limited subset of [BlockHeaderConsensusInfo] for derive_consensus_parameters()

DerivedConsensusParameters

Derived consensus parameters, to be eventually turned into OwnedBlockHeaderConsensusParameters

Enums

DeriveConsensusParametersError

Error for derive_consensus_parameters()

DeriveSuperSegmentForBlockError

Error for derive_super_segments_for_block()

ShardMembershipEntropySourceError

Error for shard_membership_entropy_source

Traits

DeriveConsensusParametersChainInfo

Chain info for derive_consensus_parameters().

DeriveSuperSegmentsForBlockChainInfo

Chain info for derive_super_segments_for_block().

ShardMembershipEntropySourceChainInfo

Chain info for shard_membership_entropy_source().

Functions

derive_consensus_parameters

derive_super_segments_for_block

Derive a super segment for a block with a specified parent block number

shard_membership_entropy_source

Find shard membership entropy for a specified block number